Add kitbox folder
